Fix tests and Refactor moving files
