+++ v1-to-v2-data-migration/helpers/types.ts
@@ -177,7 +177,10 @@ export interface HistoryItem {
 }
 
 // Resolver types
-export type ResolverFunction<T = any> = (data: T, eventType: 'birth' | 'death') => any
+export type ResolverFunction<T = any> = (
+  data: T,
+  eventType: 'birth' | 'death'
+) => any
 
 export interface ResolverMap {
   [fieldId: string]: ResolverFunction
@@ -204,7 +207,7 @@ export interface EventRegistration {
   weightAtBirth?: number
   deathDate?: string
   deathDescription?: string
-  causeOfDeathEstablished?: "true" | "false"
+  causeOfDeathEstablished?: 'true' | 'false'
   causeOfDeathMethod?: string
   mannerOfDeath?: string
 }
