+++ v1-to-v2-data-migration/tests/deno.json
@@ -15,7 +15,7 @@
     "graphql": "npm:graphql@^16.0.0"
   },
   "tasks": {
-    "migration": "deno test --allow-env --allow-run --allow-net --allow-write roundtrip.test.ts",
-    "transform": "deno test --allow-env --allow-run --allow-net --allow-write transform.test.ts"
+    "migration": "deno test --allow-env --allow-run --allow-net --allow-write ./integration/roundtrip.test.ts",
+    "transform": "deno test --allow-env --allow-run --allow-net --allow-write ./unit/*.test.ts"
   }
 }
